Move functions to utils and run black
